# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 
 * Do not edit manually.
 * Api
 * CBSE Study Guide API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Subject {
  id: number;
  name: string;
  grade: number;
  description: string;
  icon: string;
  chapterCount: number;
  color: string;
}

export interface Chapter {
  id: number;
  subjectId: number;
  title: string;
  chapterNumber: number;
  summary: string;
  /** @nullable */
  lessonStudy?: string | null;
  /** @nullable */
  conceptsJson?: string | null;
  questionCount: number;
}

export interface KeyConcept {
  id: number;
  title: string;
  explanation: string;
}

export type VisualType = (typeof VisualType)[keyof typeof VisualType];

export const VisualType = {
  geogebra: "geogebra",
  manim: "manim",
} as const;

export interface Visual {
  type: VisualType;
  id: string;
  title: string;
  thumbnail?: string;
}

export interface ChapterVideo {
  id: number;
  chapterId: number;
  youtubeVideoId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  displayOrder: number;
}

export interface ChapterDetail {
  id: number;
  subjectId: number;
  title: string;
  chapterNumber: number;
  summary: string;
  /** @nullable */
  lessonStudy?: string | null;
  /** @nullable */
  conceptsJson?: string | null;
  keyConcepts: KeyConcept[];
  questionCount: number;
  visuals?: Visual[];
  videos: ChapterVideo[];
}

export interface PracticeQuestion {
  id: number;
  chapterId: number;
  question: string;
  difficulty: string;
  type: string;
  /** @nullable */
  options?: string[] | null;
  answer: string;
  explanation: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  stepByStepSolution?: string | null;
}

export interface GradeSummary {
  grade: number;
  subjectCount: number;
  totalChapters: number;
}

export interface Dashboard {
  totalSubjects: number;
  totalChapters: number;
  totalQuestions: number;
  gradeSummaries: GradeSummary[];
  recentChapters: Chapter[];
}

export interface ProgressResponse {
  completedChapterIds: number[];
}

export interface ContentStats {
  totalChapters: number;
  chaptersWithLesson: number;
  chaptersWithQuestions: number;
  totalQuestions: number;
  /** @nullable */
  coveragePercent?: number | null;
  chaptersWithVideos: number;
  totalVideos: number;
}

export interface MarkProgressBody {
  sessionId: string;
  chapterId: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListSubjectsParams = {
  grade?: number;
};

export type ListPracticeQuestionsParams = {
  difficulty?: ListPracticeQuestionsDifficulty;
};

export type ListPracticeQuestionsDifficulty =
  (typeof ListPracticeQuestionsDifficulty)[keyof typeof ListPracticeQuestionsDifficulty];

export const ListPracticeQuestionsDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type GetProgressParams = {
  sessionId: string;
};

export type UnmarkChapterCompleteParams = {
  sessionId: string;
};