# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CBSE Study Guide API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cbse/subjects:
    get:
      operationId: listSubjects
      tags: [cbse]
      summary: List all CBSE subjects
      parameters:
        - name: grade
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
  /cbse/subjects/{subjectId}/chapters:
    get:
      operationId: listChapters
      tags: [cbse]
      summary: List chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /cbse/chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [cbse]
      summary: Get chapter details with key concepts
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /cbse/chapters/{chapterId}/questions:
    get:
      operationId: listPracticeQuestions
      tags: [cbse]
      summary: List practice questions for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
            enum: [easy, medium, hard]
      responses:
        "200":
          description: List of practice questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeQuestion"
  /cbse/chapters/{chapterId}/videos:
    get:
      operationId: listChapterVideos
      tags: [cbse]
      summary: List videos for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapter videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChapterVideo"
        "400":
          description: Invalid chapterId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /cbse/progress:
    get:
      operationId: getProgress
      tags: [cbse]
      summary: Get completed chapter IDs for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of completed chapter IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
    post:
      operationId: markChapterComplete
      tags: [cbse]
      summary: Mark a chapter as complete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkProgressBody"
      responses:
        "201":
          description: Marked complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: integer
        description:
          type: string
        icon:
          type: string
        chapterCount:
          type: integer
        color:
          type: string
      required:
        - id
        - name
        - grade
        - description
        - icon
        - chapterCount
        - color
    Chapter:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        chapterNumber:
          type: integer
        summary:
          type: string
        lessonStudy:
          type: ["string", "null"]
        conceptsJson:
          type: ["string", "null"]
        questionCount:
          type: integer
      required:
        - id
        - subjectId
        - title
        - chapterNumber
        - summary
        - questionCount
    KeyConcept:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        explanation:
          type: string
      required:
        - id
        - title
        - explanation
    ChapterDetail:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        chapterNumber:
          type: integer
        summary:
          type: string
        lessonStudy:
          type: ["string", "null"]
        conceptsJson:
          type: ["string", "null"]
        keyConcepts:
          type: array
          items:
            $ref: "#/components/schemas/KeyConcept"
        questionCount:
          type: integer
        visuals:
          type: array
          items:
            $ref: "#/components/schemas/Visual"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/ChapterVideo"
      required:
        - id
        - subjectId
        - title
        - chapterNumber
        - summary
        - keyConcepts
        - questionCount
        - videos
    Visual:
      type: object
      properties:
        type:
          type: string
          enum: [geogebra, manim]
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: string
      required:
        - type
        - id
        - title
    ChapterVideo:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        youtubeVideoId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - id
        - chapterId
        - youtubeVideoId
        - title
        - displayOrder
    PracticeQuestion:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        question:
          type: string
        difficulty:
          type: string
        type:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        answer:
          type: string
        explanation:
          type: string
        hint:
          type: ["string", "null"]
        stepByStepSolution:
          type: ["string", "null"]
      required:
        - id
        - chapterId
        - question
        - difficulty
        - type
        - answer
        - explanation
...
    ProgressResponse:
      type: object
      properties:
        completedChapterIds:
          type: array
          items:
            type: integer
      required:
        - completedChapterIds
...
    MarkProgressBody:
      type: object
      properties:
        sessionId:
          type: string
        chapterId:
          type: integer
      required:
        - sessionId
        - chapterId
...